Fix for humanize package to work in the Windows bundle

+++ setup.py
@@ -59,6 +59,9 @@ def include_requirements(target_path):
             if entry.name.startswith('.') or not entry.is_dir():
                 continue
             
+            if entry.name.startswith('humanize'):
+                continue
+
             if entry.name.endswith('.dist-info'):
                 shutil.rmtree(entry.path)
